+++ changelog.md
@@ -3,6 +3,7 @@
 ## 4.0.2
 **Fixes**
  * Add support for retrieving values from java `Map` types. These are still modeled as lists of key/value pairs.
+ * Log GraphQL query bodies. Private information or anything which is not intended to be logged should be passed as a variable as variables values are not logged.
 
 ## 4.0.1
 **Fixes**